+++ src/main/java/wooteco/subway/web/advice/SubwayControllerAdvice.java
@@ -8,7 +8,6 @@ import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 import wooteco.subway.exception.*;
-import wooteco.subway.web.member.InvalidAuthenticationException;
 
 @RestControllerAdvice
 public class SubwayControllerAdvice {
