+++ src/main/java/wooteco/subway/web/member/LoginMemberMethodArgumentResolver.java
@@ -8,6 +8,7 @@ import org.springframework.web.context.request.NativeWebRequest;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.method.support.ModelAndViewContainer;
 import wooteco.subway.domain.member.Member;
+import wooteco.subway.exception.InvalidAuthenticationException;
 import wooteco.subway.exception.NoResourceExistException;
 import wooteco.subway.service.member.MemberService;
