+++ src/main/java/wooteco/subway/web/member/interceptor/SessionInterceptor.java
@@ -3,7 +3,7 @@ package wooteco.subway.web.member.interceptor;
 import org.apache.logging.log4j.util.Strings;
 import org.springframework.stereotype.Component;
 import org.springframework.web.servlet.HandlerInterceptor;
-import wooteco.subway.web.member.InvalidAuthenticationException;
+import wooteco.subway.exception.InvalidAuthenticationException;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
