+++ src/test/java/wooteco/subway/web/member/favorite/FavoriteControllerTest.java
@@ -23,7 +23,7 @@ import wooteco.subway.domain.member.Member;
 import wooteco.subway.infra.JwtTokenProvider;
 import wooteco.subway.service.member.MemberService;
 import wooteco.subway.service.member.favorite.FavoriteService;
-import wooteco.subway.web.member.AuthorizationExtractor;
+import wooteco.subway.web.member.interceptor.AuthorizationExtractor;
 
 import java.util.Arrays;
 import java.util.List;
